+++ Runtime/Framework/CommandBus/DebugCommandCenter.cs
@@ -1,60 +1,25 @@
-using System;
 using UnityEngine;
-using UnityEngine.Events;
 
 namespace DT.UniStart {
-  [Flags]
-  public enum DebugCommandCenterMode {
-    None = 0,
-    Add = 1,
-    Push = 2,
-    All = Add | Push
-  }
+  public class DebugCommandBus : ICommandBus {
+    public string name = "DebugCommandBus";
+    readonly ICommandBus bus;
 
-  public class DebugCommandCenter : ICommandCenter {
-    public string name = "DebugCommandCenter";
-
-    readonly ICommandCenter inner;
-
-    readonly DebugCommandCenterMode mode;
-    // cache the enabled modes for performance
-    readonly bool isAddModeEnabled;
-    readonly bool isPushModeEnabled;
-
-    public DebugCommandCenter(ICommandCenter inner = null, DebugCommandCenterMode mode = DebugCommandCenterMode.Push) {
-      this.inner = inner ?? new CommandCenter();
-      this.mode = mode;
-      this.isAddModeEnabled = (this.mode & DebugCommandCenterMode.Add) == DebugCommandCenterMode.Add;
-      this.isPushModeEnabled = (this.mode & DebugCommandCenterMode.Push) == DebugCommandCenterMode.Push;
+    public DebugCommandBus(ICommandBus bus) {
+      this.bus = bus;
     }
 
     /// <summary>
-    /// Set the name of the DebugCommandCenter.
+    /// Set the name of the DebugCommandBus.
     /// </summary>
-    public DebugCommandCenter WithName(string name) {
+    public DebugCommandBus WithName(string name) {
       this.name = name;
       return this;
     }
 
-    public UnityAction Add<T>(UnityAction command) where T : ICommand {
-      this.LogAdd<T>();
-      return this.inner.Add<T>(command);
-    }
-
-    public UnityAction<T> Add<T>(UnityAction<T> command) where T : ICommand {
-      this.LogAdd<T>();
-      return this.inner.Add(command);
-    }
-
-    public void Push<T>(T arg) where T : ICommand {
-      if (this.isPushModeEnabled)
-        Debug.Log($"{this.name}.Push: command = {typeof(T)}");
-      this.inner.Push(arg);
-    }
-
-    void LogAdd<T>() {
-      if (this.isAddModeEnabled)
-        Debug.Log($"{this.name}.Add: command = {typeof(T)}");
+    public void Push<T>(T cmd) where T : ICommand {
+      Debug.Log($"{this.name}.Push: command = {typeof(T)}, parameter = {cmd}");
+      this.bus.Push(cmd);
     }
   }
 }