+++ CHANGELOG.md
@@ -8,9 +8,9 @@
 - **_Breaking Change_**: `Entry.GetContext` will search parent first instead of the root object.
 - **_Breaking Change_**: remove `InterceptEventBus` and `DelayedEventBus`.
   - Make `EventBus` methods virtual. You can override them to realize the same effect.
-- **_Breaking Change_**: remove `DelayedCommandBus`, rename `IWritableCommandBus` to `ICommandCenter`, remove `ICommandRepo.Get`, rename `CommandBus` to `CommandCenter`, rename `DebugCommandBus` to `DebugCommandCenter`.
+- **_Breaking Change_**: remove `DelayedCommandBus`, rename `IWritableCommandBus` to `ICommandCenter`, remove `ICommandRepo.Get`, rename `CommandBus` to `CommandCenter`.
 - **_Breaking Change_**: adding the same type to `CommandCenter` multi times will throw exception.
-- **_Breaking Change_**: rewrite `DebugCommandCenter`, use composition instead of inheritance just like `DebugEventBus`.
+- **_Breaking Change_**: rewrite `DebugCommandBus` to only log `Push` events.
 - Feat: add interface `IReadonlyIoC`.
 - Feat: add `ComponentExtension.GetOrAddComponent`.
 - Feat: add `IReadonlyIoC.GetEventBus`, `IReadonlyIoC.GetCommandBus`, `IIoCC.AddEventBus` and `IIoCC.AddCommandBus`.